debian: fix handling pre-release version numbers
Upstream releases beta versions with a `.beta` suffix (and similar
suffixes are likely to be expected for alpha and rc versions). This
commit adds the necessary bits to both our `watch` and `gbp.conf` files
in order to properly deal with such version numbers in conformity with
the Debian policy.